Add INSERT-CHAR-* functions, fix writing char in last pos (Issue #24)
Added INSERT-CHAR-AT-{CURSOR, POINT}, and special-cased
WRITE-CHAR-AT-{CURSOR, POINT} to handle writing at the very last
position in the window. This seems to be a "feature" of curses to bug
out when writing there. For the high-level API, we don't want any bugs.

This commit does not add string insertion functions, and does not fix
any bugs related to that.